changes fetchController register login getCardio and getFuerza
