+++ Frontend/controllers/fetchController.js
@@ -98,3 +98,40 @@
         const data=await response.json();
         return data;
     }
+    
+    //  Cuando recibe el Login ID, recoge los Datos del Usuario, luego los Cardio y Strength Exercises
+    /**
+     * Fetch Sends Phone and Password to Backend/userLogin and get Login ID to update user Status
+     *  @param {*} passwdIn Password to Attempt to Log In
+     * @param {*} phoneIn Phone to Attempt to 
+     */
+    export async function Login(passwdIn, phoneIn) {
+
+        const response = await fetch("http://localhost:3000/checkLogin", {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify({ passwd: passwdIn, phoneNumber: phoneIn })
+        })
+        const data=await response.json();
+        return data;
+    }
+
+    /**
+     * Fetch Register a new User on Backend/createUser
+     * @param {*} nombre Name to create User
+     * @param {*} apellidos Surname to create User
+     * @param {*} passwd Password to create User
+     * @param {*} phone Unique Phone to create User (Primary Key)
+     */
+    export async function addUser(nombre, apellidos, passwd, phone) {
+        console.log(nombre);
+        const requestOptions = {
+            method: "POST",
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify({ nameUser: nombre, lastNameUser: apellidos, passwd: passwd, phoneNumber: phone })
+        };
+        const response= await fetch("http://127.0.0.1:3000/addUser", requestOptions)
+        const data=await response.text();
+        return data;
+    }
+
